// src/components/molecules/foodMenu/menuCategories/sides/types.ts
export interface StyledSideItemProps {
  $border?: string
  $background?: string
  $borderRadius?: string
  $boxShadow?: string
  $display?: string
  $flexFlow?: string
  $margin?: string
  $maxWidth?: string
  $padding?: string
  $width?: string
  $pColor?: string
  $pDisplay?: string
  $pFontFamily?: string
  $pFontSize?: string
  $pFontWeight?: string
  $pJustifyContent?: string
  $pLetterSpacing?: string
  $pLineHeight?: string
  $pMargin?: string
  $pMaxWidth?: string
  $pPadding?: string
  $pTextAlign?: string
  $pTextShadow?: string
  $pTextTransform?: string
  $pWidth?: string
}

export interface SideProps {
  sideItemBackground?: string
  sideItemBorder?: string
  sideItemBorderRadius?: string
  sideItemBoxShadow?: string
  sideItemColor?: string
  sideItemDisplay?: string
  sideItemFlexFlow?: string
  sideItemFontFamily?: string
  sideItemFontSize?: string
  sideItemFontWeight?: string
  sideItemJustifyContent?: string
  sideItemLetterSpacing?: string
  sideItemLineHeight?: string
  sideItemMargin?: string
  sideItemMaxWidth?: string
  sideItemPadding?: string
  sideItemTextAlign?: string
  sideItemTextShadow?: string
  sideItemTextTransform?: string
  sideItemWidth?: string
  sides: string[]
  sidesContainerBackground?: string
  sidesContainerBorder?: string
  sidesContainerBorderRadius?: string
  sidesContainerBoxShadow?: string
  sidesContainerDisplay?: string
  sidesContainerFlexFlow?: string
  sidesContainerJustifyContent?: string
  sidesContainerOrderNumber?: string
  sidesContainerMargin?: string
  sidesContainerMaxWidth?: string
  sidesContainerPadding?: string
  sidesContainerTitle?: string
  sidesContainerTitleColor?: string
  sidesContainerTitleDisplay?: string
  sidesContainerTitleFontFamily?: string
  sidesContainerTitleFontSize?: string
  sidesContainerTitleFontWeight?: string
  sidesContainerTitleJustifyContent?: string
  sidesContainerTitleLetterSpacing?: string
  sidesContainerTitleLineHeight?: string
  sidesContainerTitleMargin?: string
  sidesContainerTitlePadding?: string
  sidesContainerTitleText?: string
  sidesContainerTitleTextAlign?: string
  sidesContainerTitleTextShadow?: string
  sidesContainerTitleTextTransform?: string
  sidesContainerWidth?: string

}